Revert "fix type bug when using GetAttrPointer"



# message auto-generated for no-merge-commit merge:
!2918 merge revert_int64 into master

Revert "fix type bug when using GetAttrPointer"

Created-by: ryan_li
Commit-by: ryan_li
Merged-by: cann-robot
Description: ## 描述
Revert "fix type bug when using GetAttrPointer"

## 关联的Issue
<!-- 如果这个PR是为了解决特定的Issue，请在这里提供Issue链接。例如：关联Issue #000-->
<!-- 如果这个PR是为了解决特定的问题单，请在这里描述问题单单号。-->

## 测试
<!--描述进行了哪些测试来验证你的改动。包括但不限于二级冒烟、算子泛化等。-->

## 文档更新
<!--如果这个PR包含文档的更新，请在这里指出。例如：更新了README.md文件。-->

## 类型标签
<!--  [x] 表示选中 -->
- [ ] 🐛 Bug 修复
- [ ] ✨ 新特性
- [ ] ⚡ 性能优化
- [ ] ♻️ 重构
- [ ] 🧪 测试
- [ ] 📦 构建/CI
- [ ] 🔧 配置变更
- [ ] 📝 文档更新
- [ ] ⬆️ 依赖升级
- [ ] 🔒 安全修复
- [ ] 🧹 代码清理
- [ ] ❓ 其他，请描述：



See merge request: cann/ops-transformer!2918

+++ mc2/3rd/mat_mul_v3/op_host/op_tiling/arch35/matmul_v3_tiling_advanced.cpp
@@ -241,7 +241,7 @@ ge::graphStatus Mc2MatMulV3Tiling::CheckArgs()
         args_.hasBias = true;
     }
     if (attrs->GetAttrNum() >= HF32_ATTR_NUM) {
-        OPS_CHECK_NULL_WITH_CONTEXT(context_, attrs->GetAttrPointer<int64_t>(HF32_ATTR_INDEX - 1));
+        OPS_CHECK_NULL_WITH_CONTEXT(context_, attrs->GetAttrPointer<int32_t>(HF32_ATTR_INDEX - 1));
         OPS_CHECK_NULL_WITH_CONTEXT(context_, attrs->GetAttrPointer<bool>(HF32_ATTR_INDEX));
     }
     OPS_CHECK_NULL_WITH_CONTEXT(context_, context_->GetOutputDesc(0));
